refactored to use abstract class for tests
