+++ gaffer-worker/src/test/java/uk/gov/gchq/gaffer/AbstractTest.java
@@ -19,12 +19,33 @@ import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.kubernetes.client.openapi.ApiClient;
+import io.kubernetes.client.openapi.apis.CustomObjectsApi;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.web.context.WebApplicationContext;
 import java.io.IOException;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
 
 
 @SpringBootTest
 public abstract class AbstractTest {
+    protected MockMvc mvc;
+    @Autowired
+    WebApplicationContext webApplicationContext;
+
+    @Autowired
+    private ApiClient apiClient;
+    protected MvcResult token;
+
+    protected static final String TEST_GRAPH_ID = "testgraphid";
+    protected static final String TEST_GRAPH_DESCRIPTION = "Test Graph Description";
 
     protected String mapToJson(final Object obj) throws JsonProcessingException {
         ObjectMapper objectMapper = new ObjectMapper();
@@ -35,5 +56,32 @@ public abstract class AbstractTest {
 
         ObjectMapper objectMapper = new ObjectMapper();
         return objectMapper.readValue(json, clazz);
+    }
+
+    @BeforeEach
+    public void setUp() throws Exception {
+
+        this.mvc = webAppContextSetup(webApplicationContext).build();
+        // Given - I have a auth token
+        final String authRequest = "{\"username\":\"javainuse\",\"password\":\"password\"}";
+        token = mvc.perform(post("/auth")
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .content(authRequest)).andReturn();
+    }
+
+    @AfterEach
+    void tearDown() {
+        CustomObjectsApi apiInstance = new CustomObjectsApi(apiClient);
+        String group = "gchq.gov.uk"; // String | the custom resource's group
+        String version = "v1"; // String | the custom resource's version
+        String namespace = "kai-helm-3"; // String | The custom resource's namespace
+        String plural = "gaffers"; // String | the custom resource's plural name. For TPRs this would be lowercase plural kind.
+        String name = TEST_GRAPH_ID; // String | the custom object's name
+
+        try{
+            apiInstance.deleteNamespacedCustomObject(group, version, namespace, plural, name, null, null, null, null, null);
+        }catch(Exception e){
+
+        }
     }
 }
